Refactor UserFormModal component and add useUserFormManagement hook
